Bugs 24, 18 - signpost dialog and walljump changes
Not too worried about infinite wall jumping, super metroid has it. I did
change the event so you can't just hold down the jump key though. Added
transparecny to the signpost dialog and made the text box bigger as well
as centering it on the signpost and resituating the text. Also the JS is
now minified.